+++ src/PolyMessage/Proxy/OperationInterceptor.cs
@@ -64,7 +64,7 @@ namespace PolyMessage.Proxy
             // the Task<> is not covariant so we cannot cast Task<object> reference to Task<Response>
             // so we do it manually but with generated code in order to avoid reflection at runtime
             Type responseType = invocation.Method.ReturnType.GenericTypeArguments[0];
-            int responseTypeID = _messageMetadata.GetMessageTypeID(responseType);
+            short responseTypeID = _messageMetadata.GetMessageTypeID(responseType);
             Task responseTask = _castDelegate(responseTypeID, responseMessage);
             invocation.ReturnValue = responseTask;
         }
